Revert "Specify SHA1 to fetch and force it as head for the specified branch."
This reverts commit a0a0e5929ad4ca2883831d5e8d5887062ee31dcd.

It looks like this is only supported in git 2.5+

Ubuntu trusty is on 1.9 (and its end of life is on April 2019).
Ubuntu xenial is on 2.7 (with an end of life of April 2021).

As a means of supporting git 1.9, the shallow fetch can look up to a few commits back in history. This functionality is only needed when a few commits are pushed to a remote branch at relatively the same time -- basically, so fast that CI cannot do an optimized checkout of the build between the two commits.

I am guessing looking back 5 commits should always suffice.

Something to note: this bug also implies that trying to run a CI build from the past will not work (if the branch HEAD is way past the requested build SHA1). However, I hope everyone is caching their git folder which would fix this problem.

Once we can utilize git 2.5+ then we can actually forget about this problem and checkout any specific SHA1!